backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
python-dotenv
cryptography
faker
watchfiles
scikit-learn
numpy